Move the test-only SetupRedis to redistest package, adjust tests
