+++ server/datastore/redis/redis_test.go
@@ -1,14 +1,11 @@
 package redis
 
 import (
-	"fmt"
 	"io"
 	"testing"
 	"time"
 
-	"github.com/fleetdm/fleet/v4/server/fleet"
-	"github.com/gomodule/redigo/redis"
-	"github.com/mna/redisc"
+	redigo "github.com/gomodule/redigo/redis"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/require"
 )
@@ -36,8 +33,8 @@ func (redisConn) Flush() error                                       { return er
 func (redisConn) Receive() (interface{}, error)                      { return nil, errFromConn }
 
 func TestConnectRetry(t *testing.T) {
-	mockDial := func(err error) func(net, addr string, opts ...redis.DialOption) (redis.Conn, error) {
-		return func(net, addr string, opts ...redis.DialOption) (redis.Conn, error) {
+	mockDial := func(err error) func(net, addr string, opts ...redigo.DialOption) (redigo.Conn, error) {
+		return func(net, addr string, opts ...redigo.DialOption) (redigo.Conn, error) {
 			var ne *netError
 			if errors.As(err, &ne) {
 				ne.countCalls++
@@ -113,141 +110,4 @@ func TestConnectRetry(t *testing.T) {
 			require.Contains(t, err.Error(), wantErr.Error())
 		})
 	}
-}
-
-func TestRedisPoolConfigureDoer(t *testing.T) {
-	const prefix = "TestRedisPoolConfigureDoer:"
-
-	t.Run("standalone", func(t *testing.T) {
-		pool := SetupRedis(t, false, false)
-
-		c1 := pool.Get()
-		defer c1.Close()
-		c2 := pool.ConfigureDoer(pool.Get())
-		defer c2.Close()
-
-		// both conns work equally well, get nil because keys do not exist,
-		// but no redirection error (this is standalone redis).
-		_, err := redis.String(c1.Do("GET", prefix+"{a}"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(c1.Do("GET", prefix+"{b}"))
-		require.Equal(t, redis.ErrNil, err)
-
-		_, err = redis.String(c2.Do("GET", prefix+"{a}"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(c2.Do("GET", prefix+"{b}"))
-		require.Equal(t, redis.ErrNil, err)
-	})
-
-	t.Run("cluster", func(t *testing.T) {
-		pool := SetupRedis(t, true, true)
-
-		c1 := pool.Get()
-		defer c1.Close()
-		c2 := pool.ConfigureDoer(pool.Get())
-		defer c2.Close()
-
-		// unconfigured conn gets MOVED error on the second key
-		// (it is bound to {a}, {b} is on a different node)
-		_, err := redis.String(c1.Do("GET", prefix+"{a}"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(c1.Do("GET", prefix+"{b}"))
-		rerr := redisc.ParseRedir(err)
-		require.Error(t, rerr)
-		require.Equal(t, "MOVED", rerr.Type)
-
-		// configured conn gets the nil value, it redirected automatically
-		_, err = redis.String(c2.Do("GET", prefix+"{a}"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(c2.Do("GET", prefix+"{b}"))
-		require.Equal(t, redis.ErrNil, err)
-	})
-}
-
-func TestEachRedisNode(t *testing.T) {
-	const prefix = "TestEachRedisNode:"
-
-	runTest := func(t *testing.T, pool fleet.RedisPool) {
-		conn := pool.Get()
-		defer conn.Close()
-		if rc, err := redisc.RetryConn(conn, 3, 100*time.Millisecond); err == nil {
-			conn = rc
-		}
-
-		for i := 0; i < 10; i++ {
-			_, err := conn.Do("SET", fmt.Sprintf("%s%d", prefix, i), i)
-			require.NoError(t, err)
-		}
-
-		var keys []string
-		err := EachRedisNode(pool, func(conn redis.Conn) error {
-			var cursor int
-			for {
-				res, err := redis.Values(conn.Do("SCAN", cursor, "MATCH", prefix+"*"))
-				if err != nil {
-					return err
-				}
-				var curKeys []string
-				if _, err = redis.Scan(res, &cursor, &curKeys); err != nil {
-					return err
-				}
-				keys = append(keys, curKeys...)
-				if cursor == 0 {
-					return nil
-				}
-			}
-		})
-		require.NoError(t, err)
-		require.Len(t, keys, 10)
-	}
-
-	t.Run("standalone", func(t *testing.T) {
-		pool := SetupRedis(t, false, false)
-		runTest(t, pool)
-	})
-
-	t.Run("cluster", func(t *testing.T) {
-		pool := SetupRedis(t, true, false)
-		runTest(t, pool)
-	})
-}
-
-func TestBindConn(t *testing.T) {
-	const prefix = "TestBindConn:"
-
-	t.Run("standalone", func(t *testing.T) {
-		pool := SetupRedis(t, false, false)
-
-		conn := pool.Get()
-		defer conn.Close()
-
-		err := BindConn(pool, conn, prefix+"a", prefix+"b", prefix+"c")
-		require.NoError(t, err)
-		_, err = redis.String(conn.Do("GET", prefix+"a"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(conn.Do("GET", prefix+"b"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(conn.Do("GET", prefix+"c"))
-		require.Equal(t, redis.ErrNil, err)
-	})
-
-	t.Run("cluster", func(t *testing.T) {
-		pool := SetupRedis(t, true, false)
-
-		conn := pool.Get()
-		defer conn.Close()
-
-		err := BindConn(pool, conn, prefix+"a", prefix+"b", prefix+"c")
-		require.Error(t, err)
-
-		err = BindConn(pool, conn, prefix+"{z}a", prefix+"{z}b", prefix+"{z}c")
-		require.NoError(t, err)
-
-		_, err = redis.String(conn.Do("GET", prefix+"{z}a"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(conn.Do("GET", prefix+"{z}b"))
-		require.Equal(t, redis.ErrNil, err)
-		_, err = redis.String(conn.Do("GET", prefix+"{z}c"))
-		require.Equal(t, redis.ErrNil, err)
-	})
 }
