+++ server/pubsub/testing_utils.go
@@ -3,12 +3,12 @@ package pubsub
 import (
 	"testing"
 
-	"github.com/fleetdm/fleet/v4/server/datastore/redis"
+	"github.com/fleetdm/fleet/v4/server/datastore/redis/redistest"
 )
 
 func SetupRedisForTest(t *testing.T, cluster bool) *redisQueryResults {
 	const dupResults = false
-	pool := redis.SetupRedis(t, cluster, false)
+	pool := redistest.SetupRedis(t, cluster, false)
 	store := NewRedisQueryResults(pool, dupResults)
 	return store
 }
