+++ server/service/async/async_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
-	"github.com/fleetdm/fleet/v4/server/datastore/redis"
+	"github.com/fleetdm/fleet/v4/server/datastore/redis/redistest"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mock"
 	redigo "github.com/gomodule/redigo/redis"
@@ -26,13 +26,13 @@ func TestCollectLabelQueryExecutions(t *testing.T) {
 
 	t.Run("standalone", func(t *testing.T) {
 		defer mysql.TruncateTables(t, ds)
-		pool := redis.SetupRedis(t, false, false)
+		pool := redistest.SetupRedis(t, false, false)
 		testCollectLabelQueryExecutions(t, ds, pool)
 	})
 
 	t.Run("cluster", func(t *testing.T) {
 		defer mysql.TruncateTables(t, ds)
-		pool := redis.SetupRedis(t, true, true)
+		pool := redistest.SetupRedis(t, true, true)
 		testCollectLabelQueryExecutions(t, ds, pool)
 	})
 }
@@ -293,13 +293,13 @@ func TestRecordLabelQueryExecutions(t *testing.T) {
 	}
 
 	t.Run("standalone", func(t *testing.T) {
-		pool := redis.SetupRedis(t, false, false)
+		pool := redistest.SetupRedis(t, false, false)
 		t.Run("sync", func(t *testing.T) { testRecordLabelQueryExecutionsSync(t, ds, pool) })
 		t.Run("async", func(t *testing.T) { testRecordLabelQueryExecutionsAsync(t, ds, pool) })
 	})
 
 	t.Run("cluster", func(t *testing.T) {
-		pool := redis.SetupRedis(t, true, false)
+		pool := redistest.SetupRedis(t, true, false)
 		t.Run("sync", func(t *testing.T) { testRecordLabelQueryExecutionsSync(t, ds, pool) })
 		t.Run("async", func(t *testing.T) { testRecordLabelQueryExecutionsAsync(t, ds, pool) })
 	})
